+++ supreSSHion/MenuBarManager.swift
@@ -51,6 +51,9 @@ class MenuBarManager: NSObject, NSApplicationDelegate, NSMenuDelegate {
         exemptionsWindow = ExemptionsWindow(supervisor: supervisor)
 
         installTerminationSignalHandlers()
+
+        NotificationPresenter.requestAuthorization()
+        NotificationPresenter.surfacePendingTerminationFailure()
     }
 
     // `killall supreSSHion` (SIGTERM) and Ctrl-C on a terminal-launched
@@ -134,6 +137,7 @@ class MenuBarManager: NSObject, NSApplicationDelegate, NSMenuDelegate {
         stateItem?.title = supervisor.supressionState.statusMessage
         keysItem?.title = supervisor.keysLoadedMessage
         keysItem?.isEnabled = supervisor.loadedKeysCount > 0
+        keysItem?.toolTip = supervisor.lastError?.localizedDescription
         resumeItem?.isHidden = !supervisor.supressionState.isDisabled
     }